日志优化

+++ frameworks/common/src/data_validator.cpp
@@ -46,7 +46,7 @@ bool DataValidator::IsDescValid(const std::string& desc)
 bool DataValidator::IsPermissionNameValid(const std::string& permissionName)
 {
     if (permissionName.empty() || (permissionName.length() > MAX_LENGTH)) {
-        ACCESSTOKEN_LOG_ERROR(LABEL, "Invalid perm length(%{public}d).", permissionName.length());
+        ACCESSTOKEN_LOG_ERROR(LABEL, "Invalid perm length(%{public}d).", static_cast<int32_t>(permissionName.length()));
         return false;
     }
     return true;
